python/pyproject.toml
[project]
name = "paleo-waifu-data"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.1.1",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
]

[tool.uv]
dev-dependencies = []
